server/requirements.txt

gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.40
glfw==2.7.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gradio==4.36.1
gradio_client==1.0.1
greenlet==3.1.1
grpcio==1.59.3
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.23.4
hyperframe==6.0.1
imageio==2.34.2
importlib-metadata==6.11.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.7.0
joblib==1.3.2
jsonpointer==2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
Levenshtein==0.25.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
maxminddb==2.6.2
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
mujoco-py==2.1.2.14
multidict==6.0.4
multiprocess==0.70.15
netifaces==0.10.6
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.54.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.5
pandas==2.1.3
pathfinding==1.0.11
Pillow==10.1.0
plotly==5.18.0
primp==0.7.0
protobuf==4.23.4
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybullet==3.2.6
pybullet-envs-gymnasium==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.6.0
Pygments==2.17.1
pyparsing==3.1.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.9.7
referencing==0.31.0
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
ruff==0.4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
simple-websocket==1.1.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.6
st-annotated-text==4.0.1
st_tailwind==1.0.5
starlette==0.37.2
streamlit==1.40.0
streamlit-chat==0.1.1
streamlit-elements==0.1.0
streamlit-image-annotation==0.4.0
streamlit-lottie==0.0.5
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tiktoken==0.7.0
tokenizer==3.4.3
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
transformers==4.41.2
transformers-stream-generator==0.0.5
treelib==1.7.0
typer==0.12.3
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.1
uvloop==0.19.0
validators==0.22.0
vector-quantize-pytorch==1.14.28
vocos==0.1.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.1
WeTextProcessing==1.0.1
wsproto==1.2.0
xpinyin==0.7.6
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
zope.event==5.0
zope.interface==7.1.1
